refactor: use host create hook
